Mexendo no ResController do User

+++ savemoneyBackend/src/main/java/com/savemoney/savemoney/Repositories/DespesaRepository.java
@@ -4,6 +4,6 @@ import com.savemoney.savemoney.Entities.Despesa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DespesaRepository extends JpaRepository<Despesa, Long>{
+public interface DespesaRepository extends JpaRepository<Despesa, Integer>{
     
 }
