+++ savemoneyBackend/src/main/java/com/savemoney/savemoney/Repositories/ReceitaRepository.java
@@ -4,7 +4,7 @@ import com.savemoney.savemoney.Entities.Receita;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceitaRepository extends JpaRepository<Receita, Long>{
+public interface ReceitaRepository extends JpaRepository<Receita, Integer>{
 
     
 }
